# Conflicts:
#	package-lock.json
#	src/index.js

+++ src/controllers/ProductController.js
@@ -7,8 +7,8 @@ const {errorResponse} = require("../lib/response");
 class ProductController {
 
   listProduct(req, res) {
-    Product.ListProduct(req.con, (err, resultProduct) => {
-      if (err) return res.status(503).json(errorResponse(503, 'Server error'));
+    Product.ListProduct(req.con,[req.query.page, req.query.size] ,(err, resultProduct) => {
+      if (err) return res.status(503).json(errorResponse(503, 'Server error',err));
       if (resultProduct) {
         Product.ListImage(req.con, (err, resultImage) => {
           if (err) return res.status(503).json(errorResponse(503, 'Server error'));
@@ -21,44 +21,15 @@ class ProductController {
                 }
               }
             }
-            return res.status(200).json(successResponse(200,{
-              products: resultProduct
-            }));
-          }
-        })
-      }
-    })
-  }
+            Product.countProduct(req.con,(err, count)=>{
+              if (err) return res.status(503).json(errorResponse(503, 'Server error'));
+              return res.status(200).json(successResponse(200,{
+                count: count[0].count,
+                products: resultProduct
+              }));
 
-  editProduct(req, res) {
-    Product.DetailProduct(req.con, req.params.id_product, (err, result) => {
-      if (err) return res.redirect('/product')
-      if (result[0]) {
-        result[0].src = [];
-        result[0].size = []
-        result[0].brand = []
-        result[0].category = []
-        result[0].seasion = []
-        Product.ListImageProduct(req.con, req.params.id_product, (err, resultImage) => {
-          if (err) return res.redirect('/product')
-          if (resultImage) {
-            for (let i in resultImage) {
-              result[0].src.push(resultImage[i].src)
-            }
-            Product.SizeProduct(req.con, req.params.id_product, (err, resultSize) => {
-              result[0].size = resultSize;
-              Product.ListOption(req.con, 'seasion', (err, resultSeasion) => {
-                result[0].seasion = resultSeasion;
-                Product.ListOption(req.con, 'brand', (err, resultBrand) => {
-                  result[0].brand = resultBrand;
-                  Product.ListOption(req.con, 'category', (err, resultCategory) => {
-                    result[0].category = resultCategory;
-                    // res.send(result[0]);
-                    res.render('editproduct', {data: result[0]})
-                  })
-                })
-              })
             })
+
           }
         })
       }
@@ -66,7 +37,6 @@ class ProductController {
   }
 
   addProductFinal(req, res) {
-
     var fileName = req.files.map(function (item, index) {
       return `uploads/` + item.filename;
     })
@@ -76,20 +46,19 @@ class ProductController {
     req.body.src = fileName;
     Product.AddProduct(req.con, req.body, (err, result) => {
       if (err){
-        return res.status(503).json(errorResponse(503, 'Server error'));
+        return res.status(503).json(errorResponse(503, 'Product error',err));
       }
       if (result) {
         var values = []
-        if (typeof req.body.size ==='object'){
-          for (let i in req.body.size) {
-            values.push([result.insertId, req.body.size[i], Number.parseInt(req.body.qnt[i])]);
-          }
-        }else {
-          values[0] = [result.insertId, req.body.size, Number.parseInt(req.body.qnt)]
+
+        const arrSize = req.body.size.split(',')
+        const arrQnt = req.body.qnt.split(',')
+        for (let i in arrSize) {
+          values.push([result.insertId, arrSize[i], Number.parseInt(arrQnt[i])]);
         }
         Product.AddSizeProduct(req.con, values, (errSize, resultSize) => {
           if (errSize){
-            return res.status(503).json(errorResponse(503, 'Server error'));
+            return res.status(503).json(errorResponse(503, 'Size error',errSize));
           }
           if (resultSize) {
             const src = [];
@@ -98,9 +67,9 @@ class ProductController {
             }
             Product.AddImageProduct(req.con, src, (errSrc, resultSrc) => {
               if (errSrc) {
-                return res.status(503).json(errorResponse(503, 'Server error'));
+                return res.status(503).json(errorResponse(503, 'Image error',errSrc));
               }
-              return res.status(201).json(errorResponse(201, 'OK'));
+              return res.status(201).json(successResponse(201, req.body));
             })
           }
         })
@@ -109,43 +78,47 @@ class ProductController {
   }
 
   editProductFinal(req, res) {
-    var fileName = req.files.map(function (item, index) {
-      return `uploads/` + item.filename;
-    })
-    var data = req.body;
-    data.id_product = req.params.id_product;
-    data.listImgRemove = data.listImgRemove.split(",");
-    data.srcImg = fileName;
-    data.size = []
-    if (typeof req.body.keyInfoProduct ==='object'){
-      for (let i in req.body.keyInfoProduct) {
-        data.size.push([Number.parseInt(data.id_product), req.body.keyInfoProduct[i], Number.parseInt(req.body.valueInfoProduct[i])]);
-      }
-    }else {
-      data.size[0] = [Number.parseInt(data.id_product), req.body.keyInfoProduct, Number.parseInt(req.body.valueInfoProduct)]
+    const data = req.body;
+    if (req.files.length > 0) {
+      data.srcImg = req.files.map(function (item, index) {
+        return `uploads/` + item.filename;
+      });
     }
-    Product.UpdateProduct(req.con, data, (err, result) => {
-      if (err) return res.send(err)
+
+    data.id_product = req.params.id_product;
+    const newSize = []
+    const arrSize = req.body.size.split(',')
+    const arQnt = req.body.qnt.split(',')
+    for (let i in arrSize) {
+      newSize.push([Number.parseInt(data.id_product), arrSize[i], Number.parseInt(arQnt[i])]);
+    }
+    data.size = newSize
+    Product.UpdateProduct(req.con, data, (errProduct, result) => {
+      if (errProduct) return res.status(503).json(errorResponse(503, 'Update Product Error', errProduct));
       if (result) {
-        Product.DeleteSize(req.con, data, (err, resultSize) => {
-          if (err) return res.send(err)
+        Product.DeleteSize(req.con, data, (errDeleteSize, resultSize) => {
+          if (errDeleteSize) return res.status(503).json(errorResponse(503, 'Delete Size Error', errDeleteSize));
           if (resultSize) {
-            Product.AddSizeProduct(req.con, data.size, (err, resultSizeAdd) => {
-                if (err) return res.send(err)
+            Product.AddSizeProduct(req.con, data.size, (errAddSize, resultSizeAdd) => {
+                if (errAddSize) return res.status(503).json(errorResponse(503, 'Add Size Error', errAddSize));
                 if (resultSizeAdd) {
-                  Product.DeleteImage(req.con, data.listImgRemove, (err, resultImageRemove) => {
-                    if (err) return res.send(err)
-                    if (resultImageRemove) {
-                      var src = []
-                      for (let i in fileName) {
-                        src.push([data.id_product, fileName[i], 'product']);
+                  if (req.files.length > 0){
+                    Product.DeleteImage(req.con, data, (errDeleteImg, resultImageRemove) => {
+                      if (errDeleteImg) return res.status(503).json(errorResponse(503, 'Delete Img Error', errDeleteImg));
+                      if (resultImageRemove) {
+                        var src = []
+                        for (let objSrc of data.srcImg) {
+                          src.push([data.id_product, objSrc, 'product']);
+                        }
+                        Product.AddImageProduct(req.con, src, (errAddImg, resultImageAdd) => {
+                          if (errAddImg) return res.status(503).json(errorResponse(503, 'Delete Img Error', errAddImg));
+                          return res.status(204).json(successResponse(204, {message: 'OK'}));
+                        })
                       }
-                      Product.AddImageProduct(req.con, src, (er, resultImageAdd) => {
-                        if (err) return res.send(err)
-                        if (resultImageRemove) res.redirect('/product');
-                      })
-                    }
-                  })
+                    })
+                  }else {
+                    return res.status(204).json(successResponse(240, {message: 'OK'}));
+                  }
                 }
               }
             )
@@ -156,21 +129,47 @@ class ProductController {
   }
 
   deleteProduct(req, res) {
-    Product.DeleteProduct(req.con, req.params.id_product, (err, result) => {
-      if (err) return res.status(503).json(errorResponse(503,'Server error'))
-      Product.DeleteSizeProduct(req.con, req.params.id_product, (err, result) => {
-        if (err) return res.status(503).json(errorResponse(503,'Server error'))
-        Product.DeleteImageProduct(req.con, req.params.id_product, (err, result) => {
-          if (err) return res.send(err)
-          return res.status(200).json(successResponse(200))
+    Product.FindProductInBill(req.con,req.params.id_product,(e,d)=>{
+      if (e) return res.status(503).json(errorResponse(503,'Server error',e))
+      if (d.length > 0) return res.status(201).json(successResponse(201,{message:'forbidden'}))
+      Product.DeleteProduct(req.con, req.params.id_product, (err, result) => {
+        if (err) return res.status(503).json(errorResponse(503,'Server error',err))
+        Product.DeleteSizeProduct(req.con, req.params.id_product, (errSize, result) => {
+          if (errSize) return res.status(503).json(errorResponse(503,'Server error',errSize))
+          Product.DeleteImageProduct(req.con, req.params.id_product, (errImg, result) => {
+            if (errImg) return res.status(503).json(errorResponse(503,'Img error',errImg))
+            return res.status(200).json(successResponse(200))
+          })
         })
       })
     })
+  }
 
+  info(req, res){
+    let data = {}
+    Product.ListCategory(req.con, (errCategory, resultCategory) => {
+      if (errCategory) return res.status(503).json(errorResponse(503,'Server error'))
+      data.category = resultCategory
+      Product.ListBrand(req.con, (errBrand, resultBrand) => {
+        if (errBrand) return res.status(503).json(errorResponse(503,'Server error'))
+        data.brand=resultBrand
+        Product.ListStyle(req.con, (errStyle, resultStyle) => {
+          if (errStyle) return res.status(503).json(errorResponse(503,'Server error'))
+          data.style=resultStyle
+          res.status(200).json(successResponse(200, data))
+        })
+      })
+    })
   }
 
   listBrand(req, res) {
     Product.ListBrand(req.con, (err, result) => {
+      if (result) return res.json(result);
+    })
+  }
+
+  listStyle(req, res) {
+    Product.ListStyle(req.con, (err, result) => {
       if (result) return res.json(result);
     })
   }
@@ -184,11 +183,6 @@ class ProductController {
   listProductApi(req, res) {
     Product.ListProduct(req.con, (err, result) => {
       if (result) return res.json(result)
-    })
-  }
-  listStyle(req, res) {
-    Product.ListStyle(req.con, (err, result) => {
-      if (result) return res.json(result);
     })
   }
